developer/django_practices/monolith/requirements.txt
Django==3.0.9
django-extra-views==0.12.0
django-pure-pagination==0.3.0
django-redis-cache==1.7.1
django-wkhtmltopdf-3.3.0
python-dotenv==0.13.0
bokeh==0.13.0
celery==4.2.1
Pillow==5.2.0
redis==2.10.6

